refac(math): correct the locations of the `Copyable` implementations
Implementations of `Copyable` that may not be applicable on curves other than
the short-weierstrass type have been moved to their appropriate locations.
