+++ tachyon/math/elliptic_curves/affine_point.h
@@ -8,8 +8,7 @@
 #include "tachyon/math/elliptic_curves/point_conversions_forward.h"
 #include "tachyon/math/geometry/point2.h"
 
-namespace tachyon {
-namespace math {
+namespace tachyon::math {
 
 template <typename Curve, typename SFINAE = void>
 class AffinePoint;
@@ -39,34 +38,6 @@ struct PointConversions<AffinePoint<SrcCurve>, AffinePoint<DstCurve>,
   }
 };
 
-}  // namespace math
-
-namespace base {
-
-template <typename Curve>
-class Copyable<math::AffinePoint<Curve>> {
- public:
-  static bool WriteTo(const math::AffinePoint<Curve>& point, Buffer* buffer) {
-    return buffer->WriteMany(point.x(), point.y(), point.infinity());
-  }
-
-  static bool ReadFrom(const Buffer& buffer, math::AffinePoint<Curve>* point) {
-    using BaseField = typename math::AffinePoint<Curve>::BaseField;
-    BaseField x, y;
-    bool infinity;
-    if (!buffer.ReadMany(&x, &y, &infinity)) return false;
-
-    *point = math::AffinePoint<Curve>(std::move(x), std::move(y), infinity);
-    return true;
-  }
-
-  static size_t EstimateSize(const math::AffinePoint<Curve>& point) {
-    return base::EstimateSize(point.x()) + base::EstimateSize(point.y()) +
-           base::EstimateSize(point.infinity());
-  }
-};
-
-}  // namespace base
-}  // namespace tachyon
+}  // namespace tachyon::math
 
 #endif  // TACHYON_MATH_ELLIPTIC_CURVES_AFFINE_POINT_H_
