+++ tachyon/math/elliptic_curves/projective_point.h
@@ -7,8 +7,7 @@
 #include "tachyon/math/elliptic_curves/point_conversions_forward.h"
 #include "tachyon/math/geometry/point3.h"
 
-namespace tachyon {
-namespace math {
+namespace tachyon::math {
 
 template <typename Curve, typename SFINAE = void>
 class ProjectivePoint;
@@ -39,36 +38,6 @@ struct PointConversions<ProjectivePoint<SrcCurve>, ProjectivePoint<DstCurve>,
   }
 };
 
-}  // namespace math
-
-namespace base {
-
-template <typename Curve>
-class Copyable<math::ProjectivePoint<Curve>> {
- public:
-  static bool WriteTo(const math::ProjectivePoint<Curve>& point,
-                      Buffer* buffer) {
-    return buffer->WriteMany(point.x(), point.y(), point.z());
-  }
-
-  static bool ReadFrom(const Buffer& buffer,
-                       math::ProjectivePoint<Curve>* point) {
-    using BaseField = typename math::ProjectivePoint<Curve>::BaseField;
-    BaseField x, y, z;
-    if (!buffer.ReadMany(&x, &y, &z)) return false;
-
-    *point =
-        math::ProjectivePoint<Curve>(std::move(x), std::move(y), std::move(z));
-    return true;
-  }
-
-  static size_t EstimateSize(const math::ProjectivePoint<Curve>& point) {
-    return base::EstimateSize(point.x()) + base::EstimateSize(point.y()) +
-           base::EstimateSize(point.z());
-  }
-};
-
-}  // namespace base
-}  // namespace tachyon
+}  // namespace tachyon::math
 
 #endif  // TACHYON_MATH_ELLIPTIC_CURVES_PROJECTIVE_POINT_H_
