+++ tachyon/math/elliptic_curves/short_weierstrass/projective_point.h
@@ -18,7 +18,8 @@
 #include "tachyon/math/elliptic_curves/projective_point.h"
 #include "tachyon/math/geometry/point3.h"
 
-namespace tachyon::math {
+namespace tachyon {
+namespace math {
 
 template <typename _Curve>
 class ProjectivePoint<
@@ -230,7 +231,39 @@ class ProjectivePoint<
   BaseField z_;
 };
 
-}  // namespace tachyon::math
+}  // namespace math
+
+namespace base {
+
+template <typename Curve>
+class Copyable<math::ProjectivePoint<
+    Curve,
+    std::enable_if_t<Curve::kType == math::CurveType::kShortWeierstrass>>> {
+ public:
+  static bool WriteTo(const math::ProjectivePoint<Curve>& point,
+                      Buffer* buffer) {
+    return buffer->WriteMany(point.x(), point.y(), point.z());
+  }
+
+  static bool ReadFrom(const Buffer& buffer,
+                       math::ProjectivePoint<Curve>* point) {
+    using BaseField = typename math::ProjectivePoint<Curve>::BaseField;
+    BaseField x, y, z;
+    if (!buffer.ReadMany(&x, &y, &z)) return false;
+
+    *point =
+        math::ProjectivePoint<Curve>(std::move(x), std::move(y), std::move(z));
+    return true;
+  }
+
+  static size_t EstimateSize(const math::ProjectivePoint<Curve>& point) {
+    return base::EstimateSize(point.x()) + base::EstimateSize(point.y()) +
+           base::EstimateSize(point.z());
+  }
+};
+
+}  // namespace base
+}  // namespace tachyon
 
 #include "tachyon/math/elliptic_curves/short_weierstrass/projective_point_impl.h"
